+++ deimos/grid.py
@@ -35,7 +35,7 @@ def data2grid(features, dims=['mz', 'drift_time', 'retention_time']):
     idx_i = [x[-1] for x in idx]
     idx = [x[0] for x in idx]
 
-    grid = np.full([len(x) for x in idx], np.nan, dtype=np.float32)
+    grid = np.full([len(x) for x in idx], np.nan, dtype=float)
     grid[tuple(idx_i)] = features.loc[:, 'intensity'].values
 
     return idx, grid
@@ -102,4 +102,4 @@ def grid2df(edges, grid, dims=['mz', 'drift_time', 'retention_time'],
 
     del additional, idx
 
-    return pd.DataFrame(np.hstack(data), columns=cols, dtype=np.float32)
+    return pd.DataFrame(np.hstack(data), columns=cols, dtype=float)
